♻️[Refactor] 컴파운드 패턴 구조 리팩토링
