+++ src/components/chat-container/chat-bubble/ChatBubble.tsx
@@ -1,236 +1,42 @@
-import { HTMLAttributes } from 'react';
-import { twMerge } from 'tailwind-merge';
-import Avatar from '@/components/avatar/Avatar';
-import { HostIcon } from '../../../../public/icons';
+import MeBubble, { MeProps } from './components/MeBubble';
+import OpponentBubble, { OpponentProps } from './components/OpponentBubble';
+import SystemBubble, { SystemProps } from './components/SystemBubble';
+import {
+  ChatBubbleContainer,
+  ChatBubbleBox,
+  ChatBubbleTime,
+  ChatBubbleProfile,
+  ChatBubbleSystem,
+} from './components/atoms/atoms';
 
-interface ChatBubbleContainerProps extends HTMLAttributes<HTMLDivElement> {
-  variant: 'ME' | 'OPPONENT' | 'SYSTEM';
-}
-
-function ChatBubbleContainer({
-  variant,
-  className,
-  children,
-  ...props
-}: ChatBubbleContainerProps) {
-  return (
-    <div
-      className={twMerge(
-        'flex w-full',
-        variant === 'ME' && 'justify-end',
-        variant === 'OPPONENT' && 'justify-start',
-        variant === 'SYSTEM' && 'justify-center',
-        className,
-      )}
-      {...props}
-    >
-      {children}
-    </div>
-  );
-}
-
-interface ChatBubbleBoxProps extends HTMLAttributes<HTMLDivElement> {
-  variant: 'ME' | 'OPPONENT';
-}
-
-function ChatBubbleBox({
-  variant,
-  children,
-  className,
-  ...props
-}: ChatBubbleBoxProps) {
-  return (
-    <div
-      className={twMerge(
-        'whitespace-pre-wrap rounded-xl px-4 py-[10px]',
-        variant === 'ME' && [
-          'rounded-tr-none bg-green-light-01 font-medium text-green-dark-03',
-        ],
-        variant === 'OPPONENT' && [
-          'rounded-tl-none bg-gray-normal-01 text-gray-darker',
-        ],
-        className,
-      )}
-      {...props}
-    >
-      {children}
-    </div>
-  );
-}
-
-interface ChatBubbleTimeProps extends HTMLAttributes<HTMLSpanElement> {}
-
-function ChatBubbleTime({
-  children,
-  className,
-  ...props
-}: ChatBubbleTimeProps) {
-  return (
-    <span
-      className={twMerge(
-        'whitespace-nowrap text-xs font-light text-gray-normal-03',
-        className,
-      )}
-      {...props}
-    >
-      {children}
-    </span>
-  );
-}
-
-interface ChatBubbleProfileProps extends HTMLAttributes<HTMLDivElement> {
-  name: string;
-  imageUrl: string;
-  isHost?: boolean;
-}
-
-function ChatBubbleProfile({
-  name,
-  imageUrl,
-  isHost,
-  className,
-  onClick,
-  ...props
-}: ChatBubbleProfileProps) {
-  return (
-    <div className={twMerge('flex items-start gap-3', className)} {...props}>
-      <div
-        className={twMerge('relative', onClick && 'cursor-pointer')}
-        onClick={onClick}
-        data-testid="profile-image"
-      >
-        <Avatar src={imageUrl} alt={name} size={'mdLg'} />
-        {isHost && (
-          <div className="absolute -right-1 -top-1 flex h-[18px] w-[18px] items-center justify-center rounded-full border-2 border-gray-normal-01 bg-green-normal-01">
-            <HostIcon width={9} height={9} data-testid="host-icon" />
-          </div>
-        )}
-      </div>
-      <div className="flex flex-col py-1.5">
-        {isHost && (
-          <span className="text-[10px] font-semibold leading-[10px] text-gray-dark-01">
-            호스트
-          </span>
-        )}
-        <span className={'font-semibold'}>{name}</span>
-      </div>
-    </div>
-  );
-}
-
-interface ChatBubbleSystemProps extends HTMLAttributes<HTMLSpanElement> {
-  username: string;
-  action: 'JOIN' | 'LEAVE';
-}
-
-function ChatBubbleSystem({
-  username,
-  action,
-  className,
-  ...props
-}: ChatBubbleSystemProps) {
-  return (
-    <span className={twMerge('text-gray-darker', className)} {...props}>
-      {`${username}님이 ${
-        action === 'JOIN' ? '들어오셨습니다.' : '나가셨습니다.'
-      }`}
-    </span>
-  );
-}
-
+// TODO: 제네릭 처리
 interface ChatBubbleComponentProps {
   variant: 'ME' | 'OPPONENT' | 'SYSTEM';
   props: MeProps | OpponentProps | SystemProps;
 }
 
-interface MeProps {
-  content: string;
-  time?: string;
-}
-
-interface OpponentProps {
-  content: string;
-  time?: string;
-  name: string;
-  profileImage: string;
-  isHost?: boolean;
-  onProfileClick?: () => void;
-  isConsecutive?: boolean;
-}
-
-interface SystemProps {
-  username: string;
-  action: 'JOIN' | 'LEAVE';
-}
-
 function ChatBubble({ variant, props }: ChatBubbleComponentProps) {
-  const renderContent = () => {
-    switch (variant) {
-      case 'ME': {
-        const { content, time } = props as MeProps;
-        return (
-          <ChatBubbleContainer variant="ME">
-            <div className="flex gap-[14px]">
-              {time && (
-                <div className="flex items-end">
-                  <ChatBubbleTime>{time}</ChatBubbleTime>
-                </div>
-              )}
-              <ChatBubbleBox variant="ME">{content}</ChatBubbleBox>
-            </div>
-          </ChatBubbleContainer>
-        );
-      }
-
-      case 'OPPONENT': {
-        const {
-          content,
-          time,
-          name,
-          profileImage,
-          isHost,
-          onProfileClick,
-          isConsecutive,
-        } = props as OpponentProps;
-        return (
-          <ChatBubbleContainer variant="OPPONENT">
-            <div className="flex flex-col">
-              {!isConsecutive && (
-                <ChatBubbleProfile
-                  name={name}
-                  imageUrl={profileImage}
-                  isHost={isHost}
-                  onClick={onProfileClick}
-                />
-              )}
-              <div className="flex gap-[14px] pl-16">
-                <ChatBubbleBox variant="OPPONENT">{content}</ChatBubbleBox>
-                {time && (
-                  <div className="flex items-end">
-                    <ChatBubbleTime>{time}</ChatBubbleTime>
-                  </div>
-                )}
-              </div>
-            </div>
-          </ChatBubbleContainer>
-        );
-      }
-
-      case 'SYSTEM': {
-        const { username, action } = props as SystemProps;
-        return (
-          <ChatBubbleContainer variant="SYSTEM">
-            <ChatBubbleSystem username={username} action={action} />
-          </ChatBubbleContainer>
-        );
-      }
-
-      default:
-        return null;
-    }
-  };
-
-  return renderContent();
+  switch (variant) {
+    case 'ME':
+      return <MeBubble {...(props as MeProps)} />;
+    case 'OPPONENT':
+      return <OpponentBubble {...(props as OpponentProps)} />;
+    case 'SYSTEM':
+      return <SystemBubble {...(props as SystemProps)} />;
+    default:
+      return (
+        // TODO: 어떤식으로 처리하면 좋을지
+        <div className="flex w-full items-center justify-center text-gray-normal-03">
+          지원하지 않는 메시지 형식입니다
+        </div>
+      );
+  }
 }
+
+ChatBubble.Container = ChatBubbleContainer;
+ChatBubble.Box = ChatBubbleBox;
+ChatBubble.Time = ChatBubbleTime;
+ChatBubble.Profile = ChatBubbleProfile;
+ChatBubble.System = ChatBubbleSystem;
 
 export default ChatBubble;
