+++ src/components/chat-container/chat-bubble/components/OpponentBubble.tsx
@@ -1,0 +1,46 @@
+import ChatBubble from '../ChatBubble';
+
+export interface OpponentProps {
+  content: string;
+  time?: string;
+  name: string;
+  profileImage: string;
+  isHost?: boolean;
+  onProfileClick?: () => void;
+  isConsecutive?: boolean;
+}
+
+function OpponentBubble({
+  content,
+  time,
+  name,
+  profileImage,
+  isHost,
+  onProfileClick,
+  isConsecutive,
+}: OpponentProps) {
+  return (
+    <ChatBubble.Container variant="OPPONENT">
+      <div className="flex flex-col">
+        {!isConsecutive && (
+          <ChatBubble.Profile
+            name={name}
+            imageUrl={profileImage}
+            isHost={isHost}
+            onClick={onProfileClick}
+          />
+        )}
+        <div className="flex gap-[14px] pl-16">
+          <ChatBubble.Box variant="OPPONENT">{content}</ChatBubble.Box>
+          {time && (
+            <div className="flex items-end">
+              <ChatBubble.Time>{time}</ChatBubble.Time>
+            </div>
+          )}
+        </div>
+      </div>
+    </ChatBubble.Container>
+  );
+}
+
+export default OpponentBubble;
