--- src/components/chat-container/chat-bubble/types/index.ts
@@ -1,2 +1,0 @@
-export * from './chatBubble';
-export * from './variantChatBubble';
